feat: response들을 인터페이스로 묶어줌 response의 종류들이 다양해지면서 다형성에 활용하기 위해서
