+++ src/main/java/http/response/JoinResponse.java
@@ -1,0 +1,11 @@
+package http.response;
+
+import java.io.IOException;
+import java.io.OutputStream;
+
+public class JoinResponse implements HttpResponse{
+    @Override
+    public void respondToRequest(String urlRequest, OutputStream out) throws IOException {
+
+    }
+}
